### 2025-06-19 #8108 @kthyng
Wait actually I didn't remember we have someone on our editorial staff who does paleo stuff! @mengqi-z Let's see if @willgearty can do this and the other paleo submission instead!
