### 2025-08-14 #8812 @editorialbot
[REVIEW]: TopoChronia: A QGIS plugin for the creation of fully quantified palaeogeographic maps
**Submitting author:** @florianfranz (<a href="http://orcid.org/0000-0001-6855-0269">Florian Franziskakis</a>)
**Repository:** https://github.com/florianfranz/topo_chronia
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
