### 2025-08-29 #8812 @johnjarmitage
I am having difficulty installing the plugin. I will struggle on, but the issue is that (1) I use linux and I do not have sudo privalages on my work computer and (2) my home computer does not have QGIS installed. Sollutions tested:

1) On the work computer, I can get around the lack of sudo by installing QGIS using mamba or conda, yet the version of QGIS is too low to be compatible with the plugin. I stopped there and switched to my personal laptop.

2) I have no installation of QGIS on my personal laptop, running Ubuntu, Noble release. The standard repository likewise has a version of QGIS that is too low. I therefore added the repository to get the latest QGIS release and proceeded to install the plugin. Now I am confronted by the lack of a dependency: pandas. The solution offered is make a manual install of pandas.

I will explore how to do this such that QGIS can find my pandas install.
