### 2025-08-29 #8812 @johnjarmitage
However, on Ubuntu with QGIS installed via `apt install` it uses a python version that is externally managed such that I cannot use pip
```terminal
$ /usr/bin/python3 -m pip install pandas
error: externally-managed-environment
```
I suspect a solution will be to get QGIS to use a specific python virtual environment. This does not look trivial.
