### 2025-08-29 #8812 @florianfranz
Hi @johnjarmitage,

Thanks for pointing out this issue. When testing it on my end, I only used it on manually installed QGIS, not `apt install`.

Instead of the -`m pip install pandas` a quick fix would be to do the following:

`sudo apt install python3-pandas
`
This would install pandas in the
If this does not work, let me know, I can alternatively add the `pandas` library to the plugin directly and make it point to that local installation, which might fix the problem as well.

Let me know, and thanks for the review !

> However, on Ubuntu with QGIS installed via `apt install` it uses a python version that is externally managed such that I cannot use pip
>
> ```
> $ /usr/bin/python3 -m pip install pandas
> error: externally-managed-environment
> ```
>
> I suspect a solution will be to get QGIS to use a specific python virtual environment. This does not look trivial.
