### 2025-09-22 #8812 @willgearty
@davenquinn @johnjarmitage any progress on your reviews of this JOSS submission?
