### 2025-11-08 florianfranz/topo_chronia#4 @davenquinn
Issues running tutorial
I have not been able to get the full tutorial to run. Some of the problems were similar to those found in #2. Phase I, the "Checks" did not seem to allow me to set an output directory properly, or write out the proper files. eventually, I was able to get past this, but my preference for output directory was ignored and all output files went to my Desktop instead.

Once there were output files on the Desktop, the algorithm did run, but it keeps failing during the "Create Node Grids" step, with a hang of QGIS requiring a Force Quit, or a total panic of the QGIS application. I cannot access logs of the actual error that precipitates the crash, as QGIS does not store
these logs in a file, but I have attached TopoChronia's own error log an OS-level crash report just in case it is useful.

[errors.log](https://github.com/user-attachments/files/23429862/errors.log)
[crash-report.log](https://github.com/user-attachments/files/23429836/crash-report.log)

I am able to run the automated tests successfully, though.

I am running a stock installation of QGIS 3.42 downloaded from the official site, on a Mac (26.0/Tahoe).
