### 2025-11-08 florianfranz/topo_chronia#4 @davenquinn
This issue was found as part of https://github.com/openjournals/joss-reviews/issues/8812#issuecomment-3411509160
