### 2025-11-08 #8812 @davenquinn
I have completed my review. Overall, I had a similar experience to @johnjarmitage on some elements of the usability of the tool. For my part, installation on QGIS 3.42 (stock version from the website) was straightforward. I was asked if Pandas should be added, and that happened automatically. However, I was not able to run the tutorial, for reasons which I documented in https://github.com/florianfranz/topo_chronia/issues/4. I was, however, able to run the automated tests, which confirmed that the tool fundamentally does what it says on the tin. Overall the tool provides a distillation of an interesting and opinionated modeling approach, which I am eager to see a software ecosystem develop around.

I have a few other comments which aren't specific issues with the tool as designed but will limit its usefulness, which the author may want to consider if desired. I don't think any of these are important enough to prevent publication, but I will enumerate them nonetheless.

- **Repository size**: I cloned the repository before I realized that I could get the plugin as a pre-compiled release. In this I noticed that the repository is over 200 MB in size, most of which is from the `.git` folder (i.e., objects that were at one point committed but have been deleted). This is a very large repository, and I would encourage the author to drop old commits or otherwise slim it down.
- **Choice of architecture limits flexibility**: While the idea of implementing this as a QGIS plugin makes sense as it allows use within a GUI and atop a pre-installed spatial stack, it also makes the system brittle. Several of the problematic steps for the tutorial involve setup that essentially configures which layers to use and how. The underlying modeling process seems to fundamentally involve transforming four inputs into a set of modeled nodes, and then interpolation atop them. Why not structure the core algorithm as a library that can be installed and used directly in a more general Python GIS environment? This could still be bundled into a QGIS plugin if desired, but it would allow more flexible use in modeling pipelines and separate algorithmic code from presentational and environment management aspects. In the current mode, this operates in sort of an "uncanny valley", functioning within QGIS but requiring complex input checking and configuration files written to the filesystem.
- **Choice of license limits flexibility**: This is sort of an elaboration of the last point, but if it were usable as a library, the algorithm described here would be very useful in broader modeling systems. However, putting this code under GPL prevents its use in non-GPL software, which would comprise almost any software that must be used within broader systems (e.g., cloud-based infrastructures). I appreciate that QGIS Plugins should be GPL to properly work with that system, but is there a chance to release the model core itself as non-GPL code, and call it from the QGIS plugin? With the current license, this model will likely remain a QGIS plugin and nothing more, no matter the scientific merit of integrating it with other software.

In short, I think this software is quite interesting and, if basic fixes are made to the tutorial, absolutely publishable in its current form. However, though it is out of scope for this review, I encourage the authors to consider their goals for how this modeling code may be useful in the future. Broadly, I think this and the broader Panalesis family of models are important and innovative, and I am eager to see them useful in ways that extend beyond a QGIS plugin.
