### 2025-11-09 florianfranz/topo_chronia#4 @florianfranz
Hi @davenquinnn,

Thanks for rhe review and for raising the issue. Looking at the errors log file, it seems that the Create Node Grid step was not successful indeed because the output file path is not found. Could you check in the "input-files.txt" and let me know if you have an output path that is defined or if it empty ?

If there is indeed an output path defined, may you please try and run again the Create Node Grid step ?

Thank you very much for your valuable feedback.
