### 2025-11-11 #8812 @willgearty
@florianfranz upon reviewing your repo's license, I noticed you are including two external Python libraries within the repository. Is there a reason you don't just have these listed within your `requirements.txt` file? Is this something that is specific to QGIS plugins?
