### 2025-11-11 #8812 @willgearty
(and yes, please feel free to add a citation to the new publication in the JOSS paper)
