### 2025-12-17 #8812 @willgearty
Just checking in, it looks like there is still some debugging going on between @davenquinn and @florianfranz in https://github.com/florianfranz/topo_chronia/issues/4? Maybe @johnjarmitage could give his two cents since it sounds like he also had some issues running the tutorial?
