### 2026-01-12 0115526
Fixed JOSS paper review bugs
