+++ functions/createnodegrid/conversions/passive_margin_wedge.py
@@ -4,24 +4,24 @@ import math
 from qgis.core import (Qgis, edit, QgsVectorLayer, QgsFeatureRequest, QgsRasterLayer, QgsMessageLog, QgsField,
                        QgsVectorFileWriter, QgsGeometry, QgsPointXY, QgsFeature, QgsProject, QgsSpatialIndex)
 from ...base_tools import BaseTools
-base_tools = BaseTools()
 
 from ..tools.passive_margin_tools import PMConversionTools
-pm_tools = PMConversionTools()
 
 from ..tools.sediments_tools import SEDConversionTools
-sed_tools = SEDConversionTools()
 
 from ..tools.feature_conversion_tools import FeatureConversionTools
-feature_conversion_tools = FeatureConversionTools()
+
 
 class PMWConversion:
-    continent_polygons_path = base_tools.get_layer_path("Continent Polygons")
-    continent_polygons_layer = QgsVectorLayer(continent_polygons_path, "Continent Polygons", 'ogr')
-    output_folder_path = base_tools.get_layer_path("Output Folder")
+    def __init__(self, base_tools: BaseTools):
+        self.base_tools = base_tools
+        self.output_folder_path = self.base_tools.get_layer_path("Output Folder")
+        self.continent_polygons_path = self.base_tools.get_layer_path("Continent Polygons")
+        self.continent_polygons_layer = QgsVectorLayer(self.continent_polygons_path, "Continent Polygons", "ogr")
+        self.pm_tools = PMConversionTools(self.base_tools)
+        self.feature_conversion_tools = FeatureConversionTools(self.base_tools)
+        self.sed_tools = SEDConversionTools(self.base_tools)
 
-    def __init__(self):
-        pass
     def passive_margin_wedge_to_nodes(self,age):
         x_min = 0
         step_length = 50
@@ -35,12 +35,12 @@ class PMWConversion:
         field_idx_xc = PM_multipoints.fields().indexOf('X_CREST')
         field_idx_zr = PM_multipoints.fields().indexOf('Z_RASTER')
         field_idx_xmax = PM_multipoints.fields().indexOf('X_MAX')
-        ridge_depth = feature_conversion_tools.get_ridge_depth(age)
+        ridge_depth = self.feature_conversion_tools.get_ridge_depth(age)
         with edit(PM_multipoints):
             for feature in PM_multipoints.getFeatures():
                 feature_age = feature.attribute("FEAT_AGE")
-                z_crest = pm_tools.crest_y_passive_margin(age, feature_age)
-                x_crest = pm_tools.crest_x_passive_margin(feature_age)
+                z_crest = self.pm_tools.crest_y_passive_margin(age, feature_age)
+                x_crest = self.pm_tools.crest_x_passive_margin(feature_age)
                 PM_multipoints.changeAttributeValue(feature.id(),field_idx_zc,float(z_crest))
                 PM_multipoints.changeAttributeValue(feature.id(),field_idx_xc,float(x_crest))
                 geom = feature.geometry()
@@ -56,7 +56,7 @@ class PMWConversion:
                     if raster_depth < -5500:
                         raster_depth = -5500
                 PM_multipoints.changeAttributeValue(feature.id(),field_idx_zr,raster_depth)
-                length = -pm_tools.wedge_x_pm_new(feature_age)* 100 #Multiply by a 100 to convert from degrees to km
+                length = -self.pm_tools.wedge_x_pm_new(feature_age)* 100 #Multiply by a 100 to convert from degrees to km
                 x_max = -length
                 PM_multipoints.changeAttributeValue(feature.id(),field_idx_xmax,x_max)
         PM_multipoints.commitChanges()
@@ -82,11 +82,11 @@ class PMWConversion:
[…]
