+++ functions/createnodegrid/tools/rift_tools.py
@@ -1,18 +1,19 @@
 from ...base_tools import BaseTools
-base_tools = BaseTools()
 
 from ..tools.feature_conversion_tools import (FeatureConversionTools)
-feature_conversion_tools = FeatureConversionTools()
+
 
 class RIBConversionTools:
-    output_folder_path = base_tools.get_layer_path("Output Folder")
     APPEARANCE = "APPEARANCE"
 
-    def __init__(self):
-        pass
+    def __init__(self, base_tools: BaseTools):
+        self.base_tools = base_tools
+        self.output_folder_path = self.base_tools.get_layer_path("Output Folder")
+        self.feature_conversion_tools = FeatureConversionTools(self.base_tools)
+
 
     def rift_profile(self, distance, crest_z, through_y, feature_age, age):
-        ridge_depth = feature_conversion_tools.get_ridge_depth(age)
+        ridge_depth = self.feature_conversion_tools.get_ridge_depth(age)
         if distance < 0:
             A = (crest_z - through_y) / (0 - (-1.5))
             B = crest_z - A * 0
@@ -23,9 +24,9 @@ class RIBConversionTools:
             continent_y = 240.38
             crest_x = 0
 
-            PCM_dist = feature_conversion_tools.PCM((flex_age * distance + feature_age), ridge_depth)
-            PCM_crest_x = feature_conversion_tools.PCM((flex_age * crest_x + feature_age), ridge_depth)
-            PCM_cont_x = feature_conversion_tools.PCM((flex_age * continent_x + feature_age), ridge_depth)
+            PCM_dist = self.feature_conversion_tools.PCM((flex_age * distance + feature_age), ridge_depth)
+            PCM_crest_x = self.feature_conversion_tools.PCM((flex_age * crest_x + feature_age), ridge_depth)
+            PCM_cont_x = self.feature_conversion_tools.PCM((flex_age * continent_x + feature_age), ridge_depth)
 
             A = (crest_z - continent_y) / (PCM_crest_x - PCM_cont_x)
             B = continent_y - A * PCM_cont_x
@@ -43,9 +44,9 @@ class RIBConversionTools:
         PARAM_RB_C1_FCURV = 1
         PARAM_RB_C1_FPCM = 500
 
-        ridge_depth = feature_conversion_tools.get_ridge_depth(age)
+        ridge_depth = self.feature_conversion_tools.get_ridge_depth(age)
 
-        crest_height_rift = feature_conversion_tools.composite(PARAM_RB_C1_mG1,
+        crest_height_rift = self.feature_conversion_tools.composite(PARAM_RB_C1_mG1,
                                       PARAM_RB_C1_sG1,
                                       PARAM_RB_C1_fG1,
                                       PARAM_RB_C1_mG2,
@@ -68,9 +69,9 @@ class RIBConversionTools:
         PARAM_RB_C2_FCURV = 1
         PARAM_RB_C2_FPCM = -500
 
-        ridge_depth = feature_conversion_tools.get_ridge_depth(age)
+        ridge_depth = self.feature_conversion_tools.get_ridge_depth(age)
 
-        through_y_rift = feature_conversion_tools.composite(PARAM_RB_C2_mG1,
+        through_y_rift = self.feature_conversion_tools.composite(PARAM_RB_C2_mG1,
                                         PARAM_RB_C2_sG1,
                                         PARAM_RB_C2_fG1,
                                         PARAM_RB_C2_mG2,
