+++ functions/createnodegrid/tools/velocity_data.py
@@ -1,4 +1,5 @@
 velocity_dict = {
+    "-5": 27.48,
     "0": 27.48,
     "1": 25.297,
     "2": 23.572,

[…]
