### 2026-01-12 4107bbf
Edited paper references
