+++ paper/paper.bib
@@ -1,179 +1,197 @@
-@article{aminovPaleogeographicReconstructionsUsing2023,
+@article{aminov2023,
   title = {Paleogeographic Reconstructions Using {{QGIS}}: {{Introducing Terra Antiqua}} Plugin and Its Application to 30 and 50 {{Ma}} Maps},
   shorttitle = {Paleogeographic Reconstructions Using {{QGIS}}},
-  author = {Aminov, Jovid and {Dupont-Nivet}, Guillaume and Ruiz, Diego and Gailleton, Boris},
-  year = {2023},
-  month = may,
-  journal = {Earth-Science Reviews},
+  author = {Aminov, Jovid and Dupont-Nivet, Guillaume and Ruiz, Diego and Gailleton, Boris},
+  date = {2023-05-01},
+  journaltitle = {Earth-Science Reviews},
+  shortjournal = {Earth-Science Reviews},
   volume = {240},
   pages = {104401},
   issn = {0012-8252},
   doi = {10.1016/j.earscirev.2023.104401},
-  urldate = {2023-10-03},
   abstract = {Paleogeographic maps play a central integrating role across all the disciplines of the Earth System Sciences. They increasingly serve as boundary conditions for climate, earth, and life models and are widely used to convey our understanding of the Earth System's evolution. We review methods and data used for making paleogeographic reconstructions that provide a background for using Terra Antiqua, a user-friendly interface software for a broader community of users and applications. The review also clarifies the context that led to the development of this new software for paleogeographic reconstructions that combines the power of GPlates, a tectonic reconstruction software, and QGIS, a leading open-source GIS software, to make the reconstruction process easy, flexible, intuitive, and fast with a set of innovative tools. Being developed as a plugin for QGIS, Terra Antiqua makes use of the QGIS Python API, providing users with advanced geospatial data processing algorithms coupled with an intuitive graphical user interface. It has a modular structure, breaking the whole reconstruction process into small steps and providing users with flexibility in their reconstructions. Terra Antiqua comes with detailed user guidelines, manuals, and video tutorials on www.paleoenvironment.eu/terra-antiqua. The current limitations and potential developments of Terra Antiqua are discussed, including, amongst others, the estimation and mapping uncertainties, generating more realistic geomorphic features, and the integration of databases in a web application.},
   keywords = {Earth System Sciences,GIS,Paleoenvironment,Paleogene,Paleogeography,Reconstruction,Software}
 }
 
-@article{belloAssessingRoleSlab2015,
+@article{bello2015,
   title = {Assessing the Role of Slab Rheology in Coupled Plate-Mantle Convection Models},
-  author = {Bello, L{\'e}a and Coltice, Nicolas and Tackley, Paul J. and Dietmar M{\"u}ller, R. and Cannon, John},
-  year = {2015},
-  month = nov,
-  journal = {Earth and Planetary Science Letters},
+  author = {Bello, Léa and Coltice, Nicolas and Tackley, Paul J. and Dietmar Müller, R. and Cannon, John},
+  date = {2015-11-15},
+  journaltitle = {Earth and Planetary Science Letters},
+  shortjournal = {Earth and Planetary Science Letters},
   volume = {430},
   pages = {191--201},
   issn = {0012-821X},
   doi = {10.1016/j.epsl.2015.08.010},
+  url = {https://www.sciencedirect.com/science/article/pii/S0012821X15005300},
   urldate = {2023-11-23},
[…]
