### 2026-01-12 49eed8a
Fixed citation in paper
