### 2026-01-12 ae26fe3
Fix Pandoc citation syntax
