### 2026-01-18 florianfranz/topo_chronia#4 @florianfranz
@davenquinn this issue is addressed in the latest release of the plugin (https://github.com/florianfranz/topo_chronia/releases/tag/v1.0.1) ,  See the recent message in https://github.com/openjournals/joss-reviews/issues/8812#issuecomment-3742730726

Can you let me know if this fixed your issue ? Many thanks !
