### 2026-02-09 #8812 @willgearty
@davenquinn any update? In the interest of moving things along, if I don't hear back from you by the end of the week, I'll take your completed checklist to mean that you are happy with the submission and it is suitable for publication.
