### 2026-02-12 #8812 @willgearty
@florianfranz could you please conduct the "Additional Author Tasks" listed above while I do some final editorial checks? Once everything looks good, I can proceed with recommending acceptance to the track editor.
