### 2026-02-12 #8812 @florianfranz
Hi @willgearty ,

- I have verified the authors affiliations and ORCIDs, all are fine.
- The latest release is v1.0.1 and is accessible here: https://github.com/florianfranz/topo_chronia/releases/tag/v1.0.1
- The release is archived on Zenodo at https://doi.org/10.5281/zenodo.18623621
- Title and authors list (including ORCIDs) in the archive match the paper
- License in the repository matches the archive (GNU General Public License v2.0 or later)

Let me know if you need anything else from my side. Thank you for everything regarding my submission !
