### 2026-02-12 #8812 @willgearty
@florianfranz I've submitted a small PR with some recommended grammar edits (a new Zenodo archive is not necessary). Also, please update the Zenodo archive to have the MIT license as well, since that is the license associated with the included geopy and GeographicLib libraries.
https://help.zenodo.org/docs/deposit/describe-records/licenses/#mixed
