### 2026-02-13 #8812 @florianfranz
Thank you @willgearty for the edits to the manuscript, I have merged your PR. Also, I have added the MIT licenses specifically for the bundled libraries. Both are now listed. Let me know if this is ok with you.
